+++ .prettierrc
@@ -1,5 +1,4 @@
 {
-  "arrowParens": "avoid",
   "semi": false,
   "singleQuote": true,
   "tabWidth": 2,
