+++ gatsby-config.js
@@ -1,13 +1,44 @@
-require('dotenv').config({
-  path: `.env.${process.env.NODE_ENV}`,
-})
-
 module.exports = {
   /* Your site config here */
   siteMetadata: {
     title: 'Intelsnap Inc.',
     description: 'Intelligence at a snap',
-    author: '@Laeeq',
+    author: 'intelsnap',
+    siteUrl: 'https://intelsnap.com',
+    topNav: [
+      {
+        name: `Home`,
+        link: `/`,
+      },
+      {
+        name: `Expertise`,
+        link: `/expertise`,
+      },
+      {
+        name: `Markets`,
+        link: `/markets`,
+      },
+      {
+        name: `Media`,
+        link: `/media`,
+      },
+      {
+        name: `About`,
+        link: `/about`,
+      },
+      {
+        name: `Contact`,
+        link: `/contact`,
+      },
+    ],
+    tools: {
+      github: '',
+      slack: '',
+      twitter: '',
+      youtube: '',
+      linkedIn: '',
+      instagram: '',
+    },
   },
   plugins: [
     `gatsby-plugin-styled-components`,
@@ -16,8 +47,8 @@ module.exports = {
     {
       resolve: `gatsby-source-filesystem`,
       options: {
-        name: `images`,
-        path: `${__dirname}/src/images/`,
+        name: `assets`,
+        path: `${__dirname}/src/assets/`,
       },
     },
   ],
