+++ src/pages/expertise.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+const Expertise = () => <Layout>Expertise placeholder</Layout>
+
+export default Expertise
