+++ src/pages/markets.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+const Markets = () => <Layout>Markets placeholder</Layout>
+
+export default Markets
