+++ src/pages/media.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+const Media = () => <Layout>Media placeholder</Layout>
+
+export default Media
